+++ browser/phasereditor2d_scripts_core/lib/phasereditor2d_scripts_core.js
@@ -108,6 +108,7 @@ class ActionTargetComp {
     }
     gameObject;
     target = "GAME_OBJECT";
+    targetName = "";
     /* START-USER-CODE */
     static getTargetGameObject(scriptNode, args) {
         const comp = ActionTargetComp.getComponent(scriptNode);
@@ -115,15 +116,46 @@ class ActionTargetComp {
             switch (comp.target) {
                 case "GAME_OBJECT":
                     return scriptNode.gameObject;
-                case "ARG_0":
-                    return args[0];
                 case "ARG_1":
-                    return args[1];
+                    return args[0];
                 case "ARG_2":
+                    return args[1];
+                case "ARG_3":
                     return args[2];
+                case "ARG_4":
+                    return args[3];
+                case "ARG_5":
+                    return args[4];
+                case "ARG_6":
+                    return args[5];
+                case "ARG_7":
+                    return args[6];
+                case "ARG_8":
+                    return args[7];
             }
         }
         return scriptNode.gameObject;
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class AddToParentActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    parentContainer;
+    /* START-USER-CODE */
+    execute(...args) {
+        const obj = this.getActionTargetObject(args);
+        this.parentContainer.add(obj);
     }
 }
 /* END OF COMPILED CODE */
@@ -144,6 +176,40 @@ class AlertActionScript extends ScriptNode {
     /* START-USER-CODE */
     execute(...args) {
         alert(this.message);
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* END-USER-IMPORTS */
+class AssignOpComp {
+    constructor(gameObject) {
+        this.gameObject = gameObject;
+        gameObject["__AssignOpComp"] = this;
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    static getComponent(gameObject) {
+        return gameObject["__AssignOpComp"];
+    }
+    gameObject;
+    operator = "=";
+    /* START-USER-CODE */
+    static computeValue(node, oldValue, newValue) {
+        const comp = AssignOpComp.getComponent(node);
+        if (comp) {
+            switch (comp.operator) {
+                case "+=":
+                    return oldValue + newValue;
+                case "*=":
+                    return oldValue * newValue;
+                default:
+                    return newValue;
+            }
+        }
+        return newValue;
     }
 }
 /* END OF COMPILED CODE */
@@ -302,6 +368,53 @@ class ExecChildrenActionScript extends ScriptNode {
 /* START OF COMPILED CODE */
 /* START-USER-IMPORTS */
 /* END-USER-IMPORTS */
+class ExecRandomActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    /* START-USER-CODE */
+    execute(...args) {
+        const child = Phaser.Utils.Array.GetRandom(this.children);
+        child.execute(...args);
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class FlipActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    horizontal = false;
+    vertical = false;
+    /* START-USER-CODE */
+    execute(...args) {
+        const obj = this.getActionTargetObject(args);
+        if (this.horizontal) {
+            obj.flipX = !obj.flipX;
+        }
+        if (this.vertical) {
+            obj.flipY = !obj.flipY;
+        }
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
 class OnAwakeScript extends ScriptNode {
     constructor(parent) {
         super(parent);
@@ -376,6 +489,14 @@ class OnEventScript extends ScriptNode {
                     });
                     break;
             }
+            // If it is attached to a game object
+            // and the emitter isn't the game object
+            // then remove the listener
+            if (this.gameObject && this.eventEmitter !== "gameObject") {
+                this.gameObject.once(Phaser.GameObjects.Events.DESTROY, () => {
+                    emitter?.off(this.eventName, this.executeChildren, this);
+                });
+            }
         }
     }
 }
@@ -404,6 +525,29 @@ class OnPointerDownScript extends OnEventScript {
             this.gameObject.setInteractive();
         }
         super.awake();
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class PlaySpriteAnimationActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    animationKey = "";
+    ignoreIfPlaying = false;
+    /* START-USER-CODE */
+    execute(...args) {
+        const obj = this.getActionTargetObject(args);
+        obj.once("animationcomplete-" + this.animationKey, () => this.executeChildren(...args));
+        obj.play(this.animationKey, this.ignoreIfPlaying);
     }
 }
 /* END OF COMPILED CODE */
@@ -460,6 +604,112 @@ class RootScript extends ScriptNode {
     }
     awake() {
         this.gameObject[`RootScript__${this.key}`] = this;
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class SetAngleActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    angle = 0;
+    /* START-USER-CODE */
+    execute(...args) {
+        const obj = this.getActionTargetObject(args);
+        obj.angle = AssignOpComp.computeValue(this, obj.angle, this.angle);
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class SetScaleXActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    scaleX = 1;
+    /* START-USER-CODE */
+    execute(...args) {
+        const obj = this.getActionTargetObject(args);
+        obj.scaleX = AssignOpComp.computeValue(this, obj.scaleX, this.scaleX);
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class SetScaleYActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    scaleY = 1;
+    /* START-USER-CODE */
+    execute(...args) {
+        const obj = this.getActionTargetObject(args);
+        obj.scaleY = AssignOpComp.computeValue(this, obj.scaleY, this.scaleY);
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class SetXActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    x = 0;
+    /* START-USER-CODE */
+    execute(...args) {
+        const obj = this.getActionTargetObject(args);
+        obj.x = AssignOpComp.computeValue(this, obj.x, this.x);
+    }
+}
+/* END OF COMPILED CODE */
+// You can write more code here
+
+// You can write more code here
+/* START OF COMPILED CODE */
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+class SetYActionScript extends ScriptNode {
+    constructor(parent) {
+        super(parent);
+        /* START-USER-CTR-CODE */
+        // Write your code here.
+        /* END-USER-CTR-CODE */
+    }
+    y = 0;
+    /* START-USER-CODE */
+    execute(...args) {
+        console.log("here");
+        const obj = this.getActionTargetObject(args);
+        obj.y = AssignOpComp.computeValue(this, obj.y, this.y);
     }
 }
 /* END OF COMPILED CODE */