+++ src/AddToParentActionScript.ts
@@ -1,0 +1,37 @@
+
+// You can write more code here
+
+/* START OF COMPILED CODE */
+
+import ScriptNode from "./ScriptNode";
+import Phaser from "phaser";
+/* START-USER-IMPORTS */
+/* END-USER-IMPORTS */
+
+export default class AddToParentActionScript extends ScriptNode {
+
+	constructor(parent: ScriptNode | Phaser.GameObjects.GameObject | Phaser.Scene) {
+		super(parent);
+
+		/* START-USER-CTR-CODE */
+		// Write your code here.
+		/* END-USER-CTR-CODE */
+	}
+
+	public parentContainer!: Phaser.GameObjects.Container | Phaser.GameObjects.Layer;
+
+	/* START-USER-CODE */
+
+	execute(...args: any[]): void {
+
+		const obj = this.getActionTargetObject(args);
+
+		this.parentContainer.add(obj);
+	}
+
+	/* END-USER-CODE */
+}
+
+/* END OF COMPILED CODE */
+
+// You can write more code here