+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## dev
+
+* The OnEventScript removes the event if it's root is a game object and the game object is destroyed.
+* Add the new components:
+    - TargetActionComp
+    - AssignOpComp.
+* Add the new scripts:
+    - AddToParentActionScript
+    - ExecRandomActionScript
+    - ExecChildrenActionScript 
+    - PlaySpriteAnimationActionScript
+    - SetAngleActionScript
+    - SetScaleXActionScript
+    - SetScaleYActionScript
+    - SetXActionScript
+    - SetYActionScript
+    - SetFlipActionScript
+
 ## v1.0.0 - Nov 17, 2023
 
 * First version.